requirements.txt
Pillow==7.0.0
Flask==2.0.1
